% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_info.R
\name{assign_info}
\alias{assign_info}
\title{Return object with simulation infos as attributes}
\usage{
assign_info(df, info)
}
\arguments{
\item{df}{Object to augment, generally a simulation data frame.}

\item{info}{A ProtoBuf message of type \code{strategic_games.Simulation.Info}.}
}
\value{
df with info attributes added.
}
\description{
Return object with simulation infos as attributes
}
