% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begriffix.R
\docType{data}
\name{begriffix}
\alias{begriffix}
\title{Begriffix data in long format}
\format{Dataset with 1304 rows and 5 variables:
\describe{
\item{y <int>}{Game score}
\item{id <chr>}{Player 1}
\item{id2 <chr>}{Player 2}
\item{j <int>}{Sequence for player 1}
\item{sp <int>}{Starting player}
}}
\usage{
data(begriffix)
}
\description{
Begriffix data in long format
}
\seealso{
Other empirical data: \code{\link{online}}
}
\concept{empirical data}
\keyword{datasets}
