% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begriffix_startbuchstaben.R
\docType{data}
\name{Begriffix_Startbuchstaben}
\alias{Begriffix_Startbuchstaben}
\title{19 manually created Begriffix Start letter combinations}
\format{Dataset with 19 rows and 4 variables where each row represents a start letter combination and each variable contains the letter for one position on the game board.
\describe{
\item{UL}{Upper left position.}
\item{UR}{Upper right position.}
\item{LL}{Lower left position.}
\item{LR}{Lower right position.}
}}
\usage{
data(Begriffix_Startbuchstaben)
}
\description{
19 manually created Begriffix Start letter combinations
}
\seealso{
Other auxiliary data: \code{\link{DERECHAR}},
  \code{\link{DEREWO_Startbuchstaben}},
  \code{\link{DEREWO_letter_pairs}}
}
\concept{auxiliary data}
\keyword{datasets}
