% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_simulation.R
\name{create_simulation}
\alias{create_simulation}
\title{Creates a simulation config from data}
\usage{
create_simulation(title, supplement = NULL, params)
}
\arguments{
\item{title}{A character scalar, a short title for the simulation.}

\item{supplement}{A character scalar, more optional explanations.}

\item{params}{A list of vectors. Each combination of these vectors becomes a simulation condition.}
}
\value{
A RProtoBuf Message of type strategic_games.Simulation.
}
\description{
A convenience helper for creating simulations with multiple conditions.
}
\details{
The params list can contain the following variables:
\describe{
\item{start_letters <chr>}{A character vector of four-letter combinations. If omitted, the dataset Begriffix_Startbuchstaben from this package is used.}
\item{trials <int>}{Indicates how many games should be played per condition.}
\item{starter_wordlist_path <chr>}{The file which contains the serialized word list for the start player.}
\item{starter_wordlist_(prefix|suffix|sample) <int>}{Indicates how many words should be taken from the starter's word list. If omitted, the complete list is used. You can supply exactly one of the three variants.}
\item{opponent_wordlist_path <chr>}{The file which contains the serialized word list for the opponent player. If omitted, starter_wordlist_path is inherited.}
\item{opponent_wordlist_(prefix|suffix|sample) <int>}{Indicates how many words should be taken from the opponent's word list. If omitted, starter_wordlist_... is used. You can supply exactly one of the three variants.}
}
}
