% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derechar.R
\docType{data}
\name{DeReChar}
\alias{DeReChar}
\title{German character frequencies}
\format{A dataset with 31 rows and 4 variables containing the german letter frequencies, one letter per row:
\describe{
\item{RF}{Relative frequency of the respective character in the corpus}
\item{AF}{Absolute frequency of the respective character in the corpus}
\item{G}{Glyph of the respective character}
\item{N}{Name of the respective character}
}}
\source{
\url{http://www1.ids-mannheim.de/fileadmin/kl/derewo/DeReChar-v-uni-030-a-l-2018-02-28-1.0.csv}
}
\usage{
data(DeReChar)
}
\description{
These frequencies were derived from the text corpus DeReKo. In this version, capitalization is not considered.
}
\references{
Korpusbasierte Zeichenhäufigkeitslisten
DERECHAR v-uni-XXX-2018-02-28-1.0,
Institut für Deutsche Sprache, Mannheim,
Februar 2018
}
\keyword{datasets}
