% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEREWO_letter_pairs.R
\docType{data}
\name{DEREWO_letter_pairs}
\alias{DEREWO_letter_pairs}
\title{Letter pairs found in DEREWO_Wortformen with their frequencies}
\format{A 1D table with 825 entries}
\usage{
data(DEREWO_letter_pairs)
}
\description{
Letter pairs found in DEREWO_Wortformen with their frequencies
}
\seealso{
Other auxiliary data: \code{\link{Begriffix_Startbuchstaben}},
  \code{\link{DERECHAR}},
  \code{\link{DEREWO_Startbuchstaben}}
}
\concept{auxiliary data}
\keyword{datasets}
