% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derewo_wortformen.R
\docType{data}
\name{DEREWO_Wortformen}
\alias{DEREWO_Wortformen}
\title{A german corpus-based ranked word list}
\format{A data frame with 100000 rows and 2 variables:
\describe{
\item{\code{word}}{character german word}
\item{\code{rank}}{integer rank in the corpus. Words with lowest rank have been found most frequently.}
}}
\source{
\url{http://www1.ids-mannheim.de/kl/projekte/methoden/derewo.html}
}
\usage{
data(DEREWO_Wortformen)
}
\description{
A dataset containing „Wortformenliste Derewo“. Please see source and reference for more informations.
}
\details{
The dataset is ordered by descending word frequency, starting with the words of highest frequency.
}
\references{
Korpusbasierte Wortformenliste DEREWO, v-100000t-2009-04-30-0.1,
mit Benutzerdokumentation,
\url{http://www.ids-mannheim.de/kl/derewo/},
© Institut für Deutsche Sprache, Programmbereich Korpuslinguistik,
Mannheim,  Deutschland, 2009.
}
\seealso{
Other word lists: \code{\link{Scrabbledict_german}},
  \code{\link{english}}
}
\concept{word lists}
\keyword{datasets}
