% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_oom.R
\docType{data}
\name{dict_oom}
\alias{dict_oom}
\title{Game duration by order of magnitude of dictionary size}
\format{Dataset with 500 rows and 2 variables:
\describe{
\item{words}{The number of words in the players' dictionaries.}
\item{moves}{How many moves did the game reach}
}}
\source{
Generated by simulation
}
\usage{
data(dict_oom)
}
\description{
This dataset contains the duration in moves of 500 simulated Begriffix games, and the corresponding number of words in the players' dictionaries. Both players shared one dictionary per game. For 5 different sizes (\code{10^(1:5)}), 100 games were generated. The start letters were randomly selected for each game from the human-generated ones. \code{\link{Derewo_Wortformen}} was used as full dictionary where the players got the nth start words.
}
\concept{simulation data}
\keyword{datasets}
