% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/english.R
\docType{data}
\name{english}
\alias{english}
\title{An english scrabble dictionary with 162825 words}
\format{A tibble with 1 variable and 162825 rows}
\usage{
data(english)
}
\description{
An english scrabble dictionary with 162825 words
}
\seealso{
Other word lists: \code{\link{DEREWO_Wortformen}},
  \code{\link{Scrabbledict_german}}
}
\concept{word lists}
\keyword{datasets}
