% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_message.R
\name{extract_message}
\alias{extract_message}
\title{Returns a RProtoBuf message as list structure}
\usage{
extract_message(x)
}
\arguments{
\item{x}{A RProtoBuf Message, list or atomic}
}
\value{
A list if x is a Message or recursive, or an atomic if x is atomic
}
\description{
\code{\link[RProtoBuf:as.list.Message]{RProtoBuf::as.list.Message()}} does not coerce the message fields. \code{\link[=extract_message]{extract_message()}} inspects the whole tree of message fields recursively and returns the message as tree structure of lists and atomic leafs.
}
