% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_boards.R
\name{initial_boards}
\alias{initial_boards}
\title{Returns initial Begriffix boards from start letters}
\usage{
initial_boards(start_letters)
}
\arguments{
\item{start_letters}{A character vector where each element contains four letters.}
}
\value{
A list of 8*8 matrices with the input start letters inserted at the middle position.
}
\description{
Returns initial Begriffix boards from start letters
}
\examples{
  sl <- c("laer", "jaul")
  initial_boards(sl)
}
