% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal}
\alias{modal}
\title{Statistical mode (central tendency measurement)}
\usage{
modal(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A scalar value with the type of x.
}
\description{
This calculates the statistical mode of an at least nominal variable, the value with the highest frequency.
}
\examples{
# metric
modal(c(1, 1, 2, 2, 2, 3, 3)) # 2
# nominal
modal(c("a", "b", "b", "c")) # b
}
