% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run a simulation config}
\usage{
run_simulation(x, path = NULL, clean = T, command = "hangman")
}
\arguments{
\item{x}{A RProtoBuf Message of type strategic_games.Simulation.}

\item{path}{The directory path where the generated .pb files are written to.}

\item{clean}{If true, the generated files are deleted after processing.}

\item{command}{The name or path to the hangman executable which must be in PATH. If omitted, ~/bin/hangman is tried.}
}
\value{
The simulation results as list tree structure.
}
\description{
This serializes the config into a .pb file, runs hangman with that file,
and reads the results produced by hangman.
}
