% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrabbledict_german.R
\docType{data}
\name{Scrabbledict_german}
\alias{Scrabbledict_german}
\title{A german scrabble dictionary with 594103 words}
\format{A tibble with 1 variable and 594103 rows}
\source{
\url{https://sourceforge.net/projects/scrabbledict/}
}
\usage{
data(Scrabbledict_german)
}
\description{
A german scrabble dictionary with 594103 words
}
\seealso{
Other word lists: \code{\link{DEREWO_Wortformen}},
  \code{\link{english}}
}
\concept{word lists}
\keyword{datasets}
