% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scrabbledict_letter_pairs.R
\docType{data}
\name{Scrabbledict_letter_pairs}
\alias{Scrabbledict_letter_pairs}
\title{Letter pairs found in Scrabbledict_german with their frequencies}
\format{A 1D table with 735 entries}
\usage{
data(Scrabbledict_letter_pairs)
}
\description{
Letter pairs found in Scrabbledict_german with their frequencies
}
\seealso{
Other letter pair frequency tables: \code{\link{DEREWO_letter_pairs}}
}
\concept{letter pair frequency tables}
\keyword{datasets}
