% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize_wordlist.R
\name{serialize_wordlist}
\alias{serialize_wordlist}
\title{Serializes a word list to a file}
\usage{
serialize_wordlist(words = NULL, source = NULL, path = NULL)
}
\arguments{
\item{words}{A character vector.}

\item{source}{A character scalar indicating the origin of the word list.}

\item{path}{The file path where to store the serialized data.}
}
\value{
The path that was used (a temp file if omitted).
}
\description{
Serializes a word list to a file
}
\examples{
library(sg.data)
words <- c("abc", "def", "xyz")
source <- "an example list"
path <- serialize_wordlist(words, source)
path
}
