% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dir.R
\docType{data}
\name{sim_dir_games}
\alias{sim_dir_games}
\alias{sim_dir_moves}
\alias{sim_dir_hits}
\title{Simulation comparing direction restrictions and strategies}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60000 rows and 7 columns.}
\source{
generated with \href{https://github.com/strategic-games/hangman}{Hangman}
}
\usage{
# Game data
data(sim_dir_games, package = "sg.data")
# Move data
data(sim_dir_moves, package = "sg.data")
# Hits data
data(sim_dir_hits, package = "sg.data")
}
\description{
This simulation contains 60,000 Begriffix games under 6 (3 first move direction restrictions x 2 common player strategies) conditions, 10,000 games per condition.
}
\details{
Except for direction restrictions, the game configuration is kept constant and conforms to original Begriffix rules.
Under each condition, the start letters are the first 10,000 entries in \code{\link{DEREWO_Startbuchstaben}}.
In each game, both players have the same strategy in common (random or min_places) and their vocabulary contains the first 50,000 words in DEREWO_Wortformen.
}
\seealso{
Other simulation data: \code{\link{sim_multifactorial_games}},
  \code{\link{sim_sl_games}}
}
\concept{simulation data}
\keyword{datasets}
