% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_multifactorial.R
\docType{data}
\name{sim_multifactorial_games}
\alias{sim_multifactorial_games}
\alias{sim_multifactorial_moves}
\alias{sim_multifactorial_hits}
\title{Simulation in multifactorial design}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36000 rows and 16 columns.}
\source{
generated with \href{https://github.com/strategic-games/hangman}{Hangman}
}
\usage{
# Game data
data(sim_multifactorial_games, package = "sg.data")
# Move data
data(sim_multifactorial_moves, package = "sg.data")
# Hits data
data(sim_multifactorial_hits, package = "sg.data")
}
\description{
This simulation contains 36,000 Begriffix games where 3,600 factor level combinations have been defined and each of them was played 10 times.
}
\seealso{
Other simulation data: \code{\link{sim_sl_games}}
}
\concept{simulation data}
\keyword{datasets}
