% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sl.R
\docType{data}
\name{sim_sl_games}
\alias{sim_sl_games}
\alias{sim_sl_moves}
\alias{sim_sl_hits}
\title{Simulation comparing start letters generation methods}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30000 rows and 6 columns.}
\source{
generated with \href{https://github.com/strategic-games/hangman}{Hangman}
}
\usage{
# Game data
data(sim_sl_games, package = "sg.data")
# Move data
data(sim_sl_moves, package = "sg.data")
# Hits data
data(sim_sl_hits, package = "sg.data")
}
\description{
This simulation contains 30,000 Begriffix games where 3 methods to generate start letter combinations are compared (10,000 games per method).
Except for start letters, the game configuration is kept constant and conforms to original Begriffix rules.
Players use the random strategy and their vocabulary contains the first 50,000 words in DEREWO_Wortformen.
}
\seealso{
Other simulation data: \code{\link{thesis_hits}},
  \code{\link{thesis_moves}}
}
\concept{simulation data}
\keyword{datasets}
