% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation2df.R
\name{simulation2df}
\alias{simulation2df}
\title{Transform a simulation to a data frame}
\usage{
simulation2df(x)
}
\arguments{
\item{x}{A list or message of type SimulationResults.}
}
\description{
Takes a list structure according to a SimulationResults message and transforms it to a tibble with nested tibbles if appropriate.
}
