% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\docType{data}
\name{simulation_strategy}
\alias{simulation_strategy}
\title{Compare players with long and short move selection strategy}
\format{Hangman simulation with 380 games.}
\source{
generated with \href{https://github.com/strategic-games/hangman}{Hangman}
}
\usage{
simulation_strategy
}
\description{
Compare players with long and short move selection strategy
}
\seealso{
Other simulation data: \code{\link{simulation_vocabulary_normal}},
  \code{\link{simulation_vocabulary}}
}
\concept{simulation data}
\keyword{datasets}
