% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\docType{data}
\name{simulation_thesis_0.6}
\alias{simulation_thesis_0.6}
\title{Full combination of simulation input variables}
\format{Hangman simulation with 10.000 games.}
\source{
generated with \href{https://github.com/strategic-games/hangman}{Hangman}
}
\usage{
simulation_thesis_0.6
}
\description{
Full combination of simulation input variables
}
\concept{simulation data}
\keyword{datasets}
