% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\docType{data}
\name{simulation_vocabulary_normal}
\alias{simulation_vocabulary_normal}
\title{Compare normally distributed vocabulary sizes (m = 50.000, s = 15.000)}
\format{Hangman simulation with 10.000 games.}
\source{
generated with \href{https://github.com/strategic-games/hangman}{Hangman}
}
\usage{
simulation_vocabulary_normal
}
\description{
Compare normally distributed vocabulary sizes (m = 50.000, s = 15.000)
}
\seealso{
Other simulation data: \code{\link{simulation_strategy}},
  \code{\link{simulation_vocabulary}}
}
\concept{simulation data}
\keyword{datasets}
