% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_hits.R
\docType{data}
\name{thesis_hits}
\alias{thesis_hits}
\title{Move hits data of 36000 simulated Begriffix games}
\format{A tibble with 6661976 observations and 6 variables:
\describe{
\item{game}{A game ID}
\item{move}{A move ID}
\item{row}{The place row.}
\item{column}{The place column.}
\item{direction}{The place direction.}
\item{words}{The words that could be used with the place.}
}}
\source{
generated with \href{https://github.com/strategic-games/hangman}{Hangman}
}
\usage{
data(thesis_hits, package="sg.data")
}
\description{
This dataset contains the hits corresponding to the moves of the 36000
simulated Begriffix games in \link{thesis_moves} in long format, where each
row represents a possible place for a certain move.
It is stored separately for quicker loading of the move data
which my thesis is mainly based on.
}
\section{Word Lists}{

Every vocabulary used in the simulated games is based on
the DeRewo Wortformenliste in \link{DEREWO_Wortformen}.
Path variables are omitted from the dataset, because they never change.
}

\seealso{
Other simulation data: \code{\link{sim_sl_games}},
  \code{\link{thesis_moves}}
}
\concept{simulation data}
\keyword{datasets}
