% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_info.R
\docType{data}
\name{thesis_info}
\alias{thesis_info}
\title{Metadata about the thesis data}
\format{A named list
\describe{
\item{title, supplement}{Textual name and description}
\item{date}{When the simulation was generated}
\item{version}{The used Hangman version}
}}
\usage{
data(thesis_info, package = "sg.data")
}
\description{
Metadata about the thesis data
}
\keyword{datasets}
