% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begriffix_startbuchstaben.R
\docType{data}
\name{Begriffix_Startbuchstaben}
\alias{Begriffix_Startbuchstaben}
\title{Begriffix Start letter combinations}
\format{A character vector}
\usage{
data(Begriffix_Startbuchstaben)
}
\description{
This set of start letter combinations was used for the empirical games.
It was not randomly created, but by humans.
}
\seealso{
Other empirical data: \code{\link{begriffix}},
  \code{\link{online}}
}
\concept{empirical data}
\keyword{datasets}
