% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_condition.R
\name{create_simulation}
\alias{create_simulation}
\title{Creates a simulation config from data}
\usage{
create_simulation(title, supplement = NULL, params)
}
\arguments{
\item{title}{A character scalar, a short title for the simulation.}

\item{supplement}{A character scalar, more optional explanations.}

\item{params}{A list of vectors.}
}
\value{
A RProtoBuf Message of type strategic_games.Simulation.
}
\description{
Creates a simulation config from data
}
