% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derechar.R
\docType{data}
\name{DERECHAR}
\alias{DERECHAR}
\title{German character frequencies}
\format{A data frame with 31 rows and 4 variables:
\describe{
\item{\code{RF}}{double relative frequency of the respective character in the corpus}
\item{\code{AF}}{double absolute frequency of the respective character in the corpus}
\item{\code{G}}{character glyph of the respective character}
\item{\code{N}}{character name of the respective character}
}}
\source{
\url{http://www1.ids-mannheim.de/fileadmin/kl/derewo/DeReChar-v-uni-030-a-l-2018-02-28-1.0.csv}
}
\usage{
data(DERECHAR)
}
\description{
This dataset contains the corpus-based list of german character frequencies (DERECHAR) reported by the Institut für deutsche Sprache in Mannheim. Each row represents one character.
}
\references{
Korpusbasierte Zeichenhäufigkeitslisten
DERECHAR v-uni-XXX-2018-02-28-1.0,
Institut für Deutsche Sprache, Mannheim,
Februar 2018
}
\seealso{
Other auxiliary data: \code{\link{Begriffix_Startbuchstaben}},
  \code{\link{DEREWO_Startbuchstaben}},
  \code{\link{DEREWO_letter_pairs}}
}
\concept{auxiliary data}
\keyword{datasets}
