% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derewo_wortformen.R
\docType{data}
\name{Derewo_Wortformen}
\alias{Derewo_Wortformen}
\title{A german corpus-based ranked word list}
\format{dataset with 100.000 rows and 2 variables: word and rank.}
\source{
\url{http://www1.ids-mannheim.de/kl/projekte/methoden/derewo.html}
}
\usage{
data(Derewo_Wortformen)
}
\description{
A dataset containing „Wortformenliste Derewo“. Please see source and reference for more informations.
}
\references{
Korpusbasierte Wortformenliste DEREWO, v-100000t-2009-04-30-0.1,
mit Benutzerdokumentation,
\url{http://www.ids-mannheim.de/kl/derewo/},
© Institut für Deutsche Sprache, Programmbereich Korpuslinguistik,
Mannheim,  Deutschland, 2009.
}
\concept{dictionaries}
\keyword{datasets}
