% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_start_letters.R
\name{generate_start_letters}
\alias{generate_start_letters}
\title{Generate a Begriffix start letter combination}
\usage{
generate_start_letters(type = "random")
}
\arguments{
\item{type}{The source type of the generated start letters (random or human).}
}
\value{
A string containing four characters.
}
\description{
Get a Begriffix start letter combination ready for passing to a simulation configuration.
}
\details{
The type parameter can take two different values:
\describe{
\item{random}{
Take a sample of size 4 from the 30 characters used to build
german words (letters, ä, ö, ü, and ß),
weighted by relative frequencies given in \code{\link{DERECHAR}},
and pastes them together as string.
}
\item{human}{
Take a random entry from \code{\link{Begriffix_Startbuchstaben}}.
}
}
}
\examples{
generate_start_letters(type = "human")
}
