% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrabbledict_german.R
\docType{data}
\name{Scrabbledict_german}
\alias{Scrabbledict_german}
\title{A scrabble dictionary with 594103 german words}
\format{A tibble with 1 variable and 594103 rows}
\source{
\url{https://sourceforge.net/projects/scrabbledict/}
}
\usage{
data(Scrabbledict_german)
}
\description{
A scrabble dictionary with 594103 german words
}
\seealso{
Other word lists: \code{\link{Derewo_Wortformen}}
}
\concept{word lists}
\keyword{datasets}
