% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\docType{data}
\name{simulation_thesis}
\alias{simulation_thesis}
\title{Full combination of simulation input variables}
\format{Hangman simulation with 10.000 games.}
\source{
generated with \href{https://github.com/strategic-games/hangman}{Hangman}
}
\usage{
simulation_thesis
}
\description{
Full combination of simulation input variables
}
\seealso{
Other simulation data: \code{\link{simulation_strategy}},
  \code{\link{simulation_vocabulary_normal}},
  \code{\link{simulation_vocabulary}}
}
\concept{simulation data}
\keyword{datasets}
