% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\docType{data}
\name{simulation_vocabulary}
\alias{simulation_vocabulary}
\title{Begriffix simulations with various vocabulary sizes}
\format{Hangman simulation with 4.750 games.}
\source{
generated with \href{https://github.com/strategic-games/hangman}{Hangman}
}
\usage{
data(simulation_vocabulary)
}
\description{
Begriffix simulations with various vocabulary sizes
}
\seealso{
Other simulation data: \code{\link{simulation_strategy}},
  \code{\link{simulation_thesis}},
  \code{\link{simulation_vocabulary_normal}}
}
\concept{simulation data}
\keyword{datasets}
